+++ src/pages/select-school/style.tsx
